Public pages

+++ src/components/profile/WarResultMap.js
@@ -1,4 +1,4 @@
-import { Link, Navigate } from "react-router-dom"
+import { Link } from "react-router-dom"
 import { deleteCalculatedRoster } from "../managers/CalculatedRosterManager"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { library } from '@fortawesome/fontawesome-svg-core'
@@ -9,6 +9,7 @@ import { faPenToSquare } from "@fortawesome/free-solid-svg-icons"
 import { useNavigate } from "react-router-dom"
 import { useContext } from "react"
 import { calculateContext } from "../views/ApplicationViews"
+import {HoverableElement }from "./HoverableElement"
 export const WarResultMap = ({ stat, getUserWarStats, setUserWarStats, publishRoster }) => {
       // Use the useContext hook to access the current value and update function of the second context
 const { setCurrentCalculateRoster } = useContext(calculateContext);
@@ -17,12 +18,18 @@ const { setCurrentCalculateRoster } = useContext(calculateContext);
  let navigate = useNavigate()
 
   const handleDeleteClick = (click, stat) => {
-    click.preventDefault()
+    click.preventDefault();
+ const userConfirmed = window.confirm(
+      "Are you sure you want to delete this roster? This action can't be undone."
+    );
+   if (userConfirmed) {
     deleteCalculatedRoster(stat.id)
     .then((res) =>  (
     getUserWarStats()
     )).then(URost => {
     setUserWarStats(URost)})
+ alert("Roster successfully deleted");
+    }
   }
   const handlePublish = (click, public_status, id) => {
     click.preventDefault();
@@ -44,25 +51,29 @@ const handleEditClick = (click) => {
 }
 
   return <div className="saved--rosters"><Link onClick={click => setCurrentCalculateRoster(stat.id)}to={`/resources/${stat.id}/view`}><div className="roster__link" to="/roster">{stat.rosterName ? `${stat.rosterName}` : "War Stats"}</div></Link>
-    <FontAwesomeIcon className="delete__roster" onClick={click => handleDeleteClick(click, stat)} icon="fa-solid fa-trash-can" />
+  
 
+<HoverableElement
+  tooltipText="edit"
+  onElementClick={handleEditClick}
+>
+  <FontAwesomeIcon
+    className="edit__roster"
+    icon="fa-solid fa-pen-to-square"
+  />
+</HoverableElement>
+
+<HoverableElement
+ tooltipText={"delete"}
+ > 
+    <FontAwesomeIcon
+     className="delete__roster" onClick={click => handleDeleteClick(click, stat)}
+      icon="fa-solid fa-trash-can" />
+      </HoverableElement>
+<HoverableElement tooltipText={"public or private"}>
 {
-  stat.is_public ? <FontAwesomeIcon className="private__roster" onClick={click => handlePublish(click, false, stat.id)} icon="fa-solid fa-rectangle-xmark" /> : <FontAwesomeIcon className="publish__roster" onClick={click => handlePublish(click, true, stat.id)} icon="fa-solid fa-share-from-square" />
-}
-
-
-    
-
-    <FontAwesomeIcon className="edit__roster"
-    
-    //navigate to edit page
-    onClick={handleEditClick}
-    icon="fa-solid fa-pen-to-square" /> 
-
-
-
-
-  </div>
-
+  stat.is_public ? <FontAwesomeIcon className="publish__roster" onClick={click => handlePublish(click, false, stat.id)} icon="fa-solid fa-rectangle-xmark" /> : <FontAwesomeIcon className="publish__roster" onClick={click => handlePublish(click, true, stat.id)} icon="fa-solid fa-share-from-square" />
+}</HoverableElement>
+</div>
 }
